Atomic design finish implementing, further configuring with best practices uphead
1. ProjectTablePage page implemented for atomic design
2. ProjectTableTemplate template implemented for atomic design
3. Provider moved from ProjectProvider to ProjectPage
