+++ src/components/molecules/icon-link/icon-link.tsx
@@ -5,16 +5,18 @@ interface IconLinkProps {
   src: string;
   alt: altType;
   className?: string;
+  onClick?: () => void;
 }
 
 export const IconLink: React.FC<IconLinkProps> = ({
   src,
   alt,
   className,
+  onClick,
   ...props
 }) => {
   return (
-    <Link className={className} {...props}>
+    <Link className={className} onClick={onClick} {...props}>
       <Icon src={src} alt={alt} />
     </Link>
   );
