--- src/features/projects/providers/projects-table-provider.tsx
@@ -1,15 +1,0 @@
-"use client";
-
-import { Provider } from "react-redux";
-import { store } from "@/redux/store";
-import ProjectsTable from "../../../components/templates/project-table/projects-table";
-
-function ProjectsTableProvider() {
-  return (
-    <Provider store={store}>
-      <ProjectsTable />
-    </Provider>
-  );
-}
-
-export default ProjectsTableProvider;
